requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.3.0
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
bce-python-sdk==0.9.6
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.51.0
frozenlist==1.4.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
loguru==0.7.2
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.8.4
multidict==6.0.5
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.2.2
pdf2docx==0.5.8
pillow==10.3.0
playwright==1.43.0
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.8
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pyee==11.1.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
rarfile==4.2
requests==2.31.0
scikit-image==0.23.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
termcolor==2.4.0
tifffile==2024.4.18
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
visualdl==2.5.3
Werkzeug==3.0.2
win32-setctime==1.1.0
yarl==1.9.4
python-dotenv~=1.0.1